back/requirements.txt
fastapi[all]
python-dotenv
SQLAlchemy
passlib
bcrypt==3.2.2
psycopg2
alembic
python-jose[cryptography]
python-decouple
